fix: Go generics cannot find common package object type definition
- use ast.SelectorExpr instead if ast.Ident to embed types
- tests reworked

fixes https://github.com/swaggo/swag/issues/1237
