+++ testdata/generics_basic/api/api.go
@@ -3,6 +3,7 @@ package api
 import (
 	"net/http"
 
+	"github.com/swaggo/swag/testdata/generics_basic/types"
 	"github.com/swaggo/swag/testdata/generics_basic/web"
 )
 
@@ -10,39 +11,39 @@ import (
 // @Description get string by ID
 // @Accept  json
 // @Produce  json
-// @Param   data        body   web.GenericBody[web.Post]    true  "Some ID"
-// @Success 200 {object} web.GenericResponse[web.Post]
-// @Success 222 {object} web.GenericResponseMulti[web.Post, web.Post]
+// @Param   data        body   web.GenericBody[types.Post]    true  "Some ID"
+// @Success 200 {object} web.GenericResponse[types.Post]
+// @Success 222 {object} web.GenericResponseMulti[types.Post, types.Post]
 // @Failure 400 {object} web.APIError "We need ID!!"
 // @Failure 404 {object} web.APIError "Can not find ID"
 // @Router /posts/ [post]
 func GetPost(w http.ResponseWriter, r *http.Request) {
 	//write your code
-	_ = web.GenericResponse[web.Post]{}
+	_ = web.GenericResponse[types.Post]{}
 }
 
 // @Summary Add new pets to the store
 // @Description get string by ID
 // @Accept  json
 // @Produce  json
-// @Param   data	body	web.GenericBodyMulti[web.Post, web.Post]	true	"Some ID"
-// @Success 200 {object} web.GenericResponse[web.Post]
-// @Success 222 {object} web.GenericResponseMulti[web.Post, web.Post]
+// @Param   data	body	web.GenericBodyMulti[types.Post, types.Post]	true	"Some ID"
+// @Success 200 {object} web.GenericResponse[types.Post]
+// @Success 222 {object} web.GenericResponseMulti[types.Post, types.Post]
 // @Router /posts-multi/ [post]
 func GetPostMulti(w http.ResponseWriter, r *http.Request) {
 	//write your code
-	_ = web.GenericResponse[web.Post]{}
+	_ = web.GenericResponse[types.Post]{}
 }
 
 // @Summary Add new pets to the store
 // @Description get string by ID
 // @Accept  json
 // @Produce  json
-// @Param   data	body	web.GenericBodyMulti[[]web.Post, [][]web.Post]	true	"Some ID"
-// @Success 200 {object} web.GenericResponse[[]web.Post]
-// @Success 222 {object} web.GenericResponseMulti[[]web.Post, [][]web.Post]
+// @Param   data	body	web.GenericBodyMulti[[]types.Post, [][]types.Post]	true	"Some ID"
+// @Success 200 {object} web.GenericResponse[[]types.Post]
+// @Success 222 {object} web.GenericResponseMulti[[]types.Post, [][]types.Post]
 // @Router /posts-multis/ [post]
 func GetPostArray(w http.ResponseWriter, r *http.Request) {
 	//write your code
-	_ = web.GenericResponse[web.Post]{}
+	_ = web.GenericResponse[types.Post]{}
 }
