+++ testdata/generics_nested/api/api.go
@@ -3,6 +3,7 @@ package api
 import (
 	"net/http"
 
+	"github.com/swaggo/swag/testdata/generics_nested/types"
 	"github.com/swaggo/swag/testdata/generics_nested/web"
 )
 
@@ -10,30 +11,30 @@ import (
 // @Description Get All of the Posts
 // @Accept  json
 // @Produce  json
-// @Param   data	body	web.GenericNestedBody[web.GenericInnerType[web.Post]]	true	"Some ID"
-// @Success 200 {object} web.GenericNestedResponse[web.Post]
-// @Success 201 {object} web.GenericNestedResponse[web.GenericInnerType[web.Post]]
-// @Success 202 {object} web.GenericNestedResponseMulti[web.Post, web.GenericInnerMultiType[web.Post, web.Post]]
-// @Success 203 {object} web.GenericNestedResponseMulti[web.Post, web.GenericInnerMultiType[web.Post, web.GenericInnerType[web.Post]]]
-// @Success 222 {object} web.GenericNestedResponseMulti[web.GenericInnerType[web.Post], web.Post]
+// @Param   data	body	web.GenericNestedBody[web.GenericInnerType[types.Post]]	true	"Some ID"
+// @Success 200 {object} web.GenericNestedResponse[types.Post]
+// @Success 201 {object} web.GenericNestedResponse[web.GenericInnerType[types.Post]]
+// @Success 202 {object} web.GenericNestedResponseMulti[types.Post, web.GenericInnerMultiType[types.Post, types.Post]]
+// @Success 203 {object} web.GenericNestedResponseMulti[types.Post, web.GenericInnerMultiType[types.Post, web.GenericInnerType[types.Post]]]
+// @Success 222 {object} web.GenericNestedResponseMulti[web.GenericInnerType[types.Post], types.Post]
 // @Router /posts [get]
 func GetPosts(w http.ResponseWriter, r *http.Request) {
-	_ = web.GenericNestedResponse[web.Post]{}
+	_ = web.GenericNestedResponse[types.Post]{}
 }
 
 // @Summary List Posts
 // @Description Get All of the Posts
 // @Accept  json
 // @Produce  json
-// @Param   data	body	web.GenericNestedBody[web.GenericInnerType[[]web.Post]]	true	"Some ID"
-// @Success 200 {object} web.GenericNestedResponse[[]web.Post]
-// @Success 201 {object} web.GenericNestedResponse[[]web.GenericInnerType[web.Post]]
-// @Success 202 {object} web.GenericNestedResponse[[]web.GenericInnerType[[]web.Post]]
-// @Success 203 {object} web.GenericNestedResponseMulti[[]web.Post, web.GenericInnerMultiType[[]web.Post, web.Post]]
-// @Success 204 {object} web.GenericNestedResponseMulti[[]web.Post, []web.GenericInnerMultiType[[]web.Post, web.Post]]
-// @Success 205 {object} web.GenericNestedResponseMulti[web.Post, web.GenericInnerMultiType[web.Post, []web.GenericInnerType[[][]web.Post]]]
-// @Success 222 {object} web.GenericNestedResponseMulti[web.GenericInnerType[[]web.Post], []web.Post]
+// @Param   data	body	web.GenericNestedBody[web.GenericInnerType[[]types.Post]]	true	"Some ID"
+// @Success 200 {object} web.GenericNestedResponse[[]types.Post]
+// @Success 201 {object} web.GenericNestedResponse[[]web.GenericInnerType[types.Post]]
+// @Success 202 {object} web.GenericNestedResponse[[]web.GenericInnerType[[]types.Post]]
+// @Success 203 {object} web.GenericNestedResponseMulti[[]types.Post, web.GenericInnerMultiType[[]types.Post, types.Post]]
+// @Success 204 {object} web.GenericNestedResponseMulti[[]types.Post, []web.GenericInnerMultiType[[]types.Post, types.Post]]
+// @Success 205 {object} web.GenericNestedResponseMulti[types.Post, web.GenericInnerMultiType[types.Post, []web.GenericInnerType[[][]types.Post]]]
+// @Success 222 {object} web.GenericNestedResponseMulti[web.GenericInnerType[[]types.Post], []types.Post]
 // @Router /posts-multis/ [get]
 func GetPostArray(w http.ResponseWriter, r *http.Request) {
-	_ = web.GenericNestedResponse[web.Post]{}
+	_ = web.GenericNestedResponse[types.Post]{}
 }
