+++ testdata/generics_property/expected.json
@@ -70,64 +70,45 @@
         }
     },
     "definitions": {
+        "types.Post": {
+            "type": "object",
+            "properties": {
+                "@uri": {
+                    "type": "string"
+                },
+                "data": {
+                    "description": "Post data",
+                    "type": "object",
+                    "properties": {
+                        "name": {
+                            "description": "Post tag",
+                            "type": "array",
+                            "items": {
+                                "type": "string"
+                            }
+                        }
+                    }
+                },
+                "id": {
+                    "type": "integer",
+                    "format": "int64",
+                    "example": 1
+                },
+                "name": {
+                    "description": "Post name",
+                    "type": "string",
+                    "example": "poti"
+                }
+            }
+        },
         "web.PostResponse": {
             "type": "object",
             "properties": {
                 "items": {
-                    "type": "object",
-                    "properties": {
-                        "data": {
-                            "description": "Post data",
-                            "type": "object",
-                            "properties": {
-                                "name": {
-                                    "description": "Post tag",
-                                    "type": "array",
-                                    "items": {
-                                        "type": "string"
-                                    }
-                                }
-                            }
-                        },
-                        "id": {
-                            "type": "integer",
-                            "format": "int64",
-                            "example": 1
-                        },
-                        "name": {
-                            "description": "Post name",
-                            "type": "string",
-                            "example": "poti"
-                        }
-                    }
+                    "$ref": "#/definitions/types.Post"
                 },
                 "items2": {
-                    "type": "object",
-                    "properties": {
-                        "data": {
-                            "description": "Post data",
-                            "type": "object",
-                            "properties": {
-                                "name": {
-                                    "description": "Post tag",
-                                    "type": "array",
-                                    "items": {
-                                        "type": "string"
-                                    }
-                                }
-                            }
-                        },
-                        "id": {
-                            "type": "integer",
-                            "format": "int64",
-                            "example": 1
-                        },
-                        "name": {
-                            "description": "Post name",
-                            "type": "string",
-                            "example": "poti"
-                        }
-                    }
+                    "$ref": "#/definitions/types.Post"
                 }
             }
         },
@@ -137,61 +118,11 @@
                 "items": {
                     "type": "array",
                     "items": {
-                        "type": "object",
-                        "properties": {
-                            "data": {
-                                "description": "Post data",
-                                "type": "object",
-                                "properties": {
-                                    "name": {
-                                        "description": "Post tag",
-                                        "type": "array",
-                                        "items": {
-                                            "type": "string"
-                                        }
-                                    }
-                                }
-                            },
-                            "id": {
-                                "type": "integer",
-                                "format": "int64",
-                                "example": 1
-                            },
-                            "name": {
-                                "description": "Post name",
-                                "type": "string",
-                                "example": "poti"
-                            }
-                        }
+                        "$ref": "#/definitions/types.Post"
                     }
                 },
                 "items2": {
-                    "type": "object",
-                    "properties": {
-                        "data": {
-                            "description": "Post data",
-                            "type": "object",
-                            "properties": {
-                                "name": {
-                                    "description": "Post tag",
-                                    "type": "array",
-                                    "items": {
-                                        "type": "string"
-                                    }
-                                }
-                            }
-                        },
-                        "id": {
-                            "type": "integer",
-                            "format": "int64",
-                            "example": 1
-                        },
-                        "name": {
-                            "description": "Post name",
-                            "type": "string",
-                            "example": "poti"
-                        }
-                    }
+                    "$ref": "#/definitions/types.Post"
                 }
             }
         },
